+++ src/simulation/call_ops.h
@@ -8,10 +8,10 @@
 class CallOps {
 public:
 
-    static bool RJMP(Environ& env, uint32_t instrn);
+    static bool RJMP(Environ& env, uint16_t instrn);
     static bool RJMP(Environ& env, TwelveBit tgtOffset, bool dummy);
 
-    static bool JMP(Environ& env, uint64_t instrn);
+    static bool JMP(Environ& env, uint32_t instrn);
     static bool JMP(Environ &env, TwentyTwoBit instrn, bool dummy);
 
     static bool IJMP(Environ& env);
@@ -20,16 +20,16 @@ public:
     static bool EIJMP(Environ& env);
     static bool EIJMP(Environ& env, bool dummy);
 
-    static bool PUSH(Environ& env, uint32_t instrn);
+    static bool PUSH(Environ& env, uint16_t instrn);
     static bool PUSH(Environ& env, FiveBit regAddr, bool dummy);
 
-    static bool POP(Environ& env, uint32_t instrn);
+    static bool POP(Environ& env, uint16_t instrn);
     static bool POP(Environ& env, FiveBit regAddr, bool dummy);
 
-    static bool RCALL(Environ& env, uint32_t instrn);
+    static bool RCALL(Environ& env, uint16_t instrn);
     static bool RCALL(Environ& env, TwelveBit tgtOffset, bool dummy);
 
-    static bool CALL(Environ& env, uint64_t instrn);
+    static bool CALL(Environ& env, uint32_t instrn);
     static bool CALL(Environ &env, TwentyTwoBit instrn, bool dummy);
 
     static bool ICALL(Environ& env);
