+++ src/simulation/sreg_ops.cpp
@@ -36,7 +36,7 @@ bool SRegOps::WriteBit(Environ &env, ThreeBit regNum, bool value) {
     return false;
 }
 
-bool SRegOps::BSET(Environ &env, u_int32_t instrn) {
+bool SRegOps::BSET(Environ &env, uint16_t instrn) {
     ThreeBit regNum;
     ArgsDecode::SBit3(instrn, regNum);
     return BSET(env, regNum, true);
@@ -46,7 +46,7 @@ bool SRegOps::BSET(Environ &env, ThreeBit regNum, bool dummy) {
     return WriteBit(env, regNum, true);
 }
 
-bool SRegOps::BCLR(Environ &env, u_int32_t instrn) {
+bool SRegOps::BCLR(Environ &env, uint16_t instrn) {
     ThreeBit regNum;
     ArgsDecode::SBit3(instrn, regNum);
     return BCLR(env, regNum, true);
@@ -56,7 +56,7 @@ bool SRegOps::BCLR(Environ &env, ThreeBit regNum, bool dummy) {
     return WriteBit(env, regNum, false);
 }
 
-bool SRegOps::SEC(Environ &env, u_int32_t instrn) {
+bool SRegOps::SEC(Environ &env, uint16_t instrn) {
     return SEC(env);
 }
 
@@ -64,7 +64,7 @@ bool SRegOps::SEC(Environ &env) {
     return BSET(env, CBit);
 }
 
-bool SRegOps::CLC(Environ &env, u_int32_t instrn) {
+bool SRegOps::CLC(Environ &env, uint16_t instrn) {
     return CLC(env);
 }
 
@@ -72,7 +72,7 @@ bool SRegOps::CLC(Environ &env) {
     return BCLR(env, CBit);
 }
 
-bool SRegOps::SEZ(Environ &env, u_int32_t instrn) {
+bool SRegOps::SEZ(Environ &env, uint16_t instrn) {
     return SEZ(env);
 }
 
@@ -80,7 +80,7 @@ bool SRegOps::SEZ(Environ &env) {
     return BSET(env, ZBit);
 }
 
-bool SRegOps::CLZ(Environ &env, u_int32_t instrn) {
+bool SRegOps::CLZ(Environ &env, uint16_t instrn) {
     return CLZ(env);
 }
 
@@ -88,7 +88,7 @@ bool SRegOps::CLZ(Environ &env) {
     return BCLR(env, ZBit);
 }
 
-bool SRegOps::SEN(Environ &env, u_int32_t instrn) {
+bool SRegOps::SEN(Environ &env, uint16_t instrn) {
     return SEN(env);
 }
 
@@ -96,7 +96,7 @@ bool SRegOps::SEN(Environ &env) {
     return BSET(env, NBit);
 }
 
-bool SRegOps::CLN(Environ &env, u_int32_t instrn) {
+bool SRegOps::CLN(Environ &env, uint16_t instrn) {
     return CLN(env);
 }
 
@@ -104,7 +104,7 @@ bool SRegOps::CLN(Environ &env) {
     return BCLR(env, NBit);
 }
 
-bool SRegOps::SEV(Environ &env, uint32_t instrn) {
+bool SRegOps::SEV(Environ &env, uint16_t instrn) {
     return SEV(env);
 }
 
@@ -112,7 +112,7 @@ bool SRegOps::SEV(Environ &env) {
     return BSET(env, VBit);
 }
 
-bool SRegOps::CLV(Environ &env, uint32_t instrn) {
+bool SRegOps::CLV(Environ &env, uint16_t instrn) {
     return CLV(env);
 }
 
@@ -120,7 +120,7 @@ bool SRegOps::CLV(Environ &env) {
     return BCLR(env, VBit);
 }
 
-bool SRegOps::SES(Environ &env, uint32_t instrn) {
+bool SRegOps::SES(Environ &env, uint16_t instrn) {
     return SES(env);
 }
 
@@ -128,7 +128,7 @@ bool SRegOps::SES(Environ &env) {
     return BSET(env, SBit);
 }
 
-bool SRegOps::CLS(Environ &env, uint32_t instrn) {
+bool SRegOps::CLS(Environ &env, uint16_t instrn) {
     return CLS(env);
 }
 
@@ -136,7 +136,7 @@ bool SRegOps::CLS(Environ &env) {
     return BCLR(env, SBit);
 }
 
-bool SRegOps::SEH(Environ &env, uint32_t instrn) {
+bool SRegOps::SEH(Environ &env, uint16_t instrn) {
     return SEH(env);
 }
 
@@ -144,7 +144,7 @@ bool SRegOps::SEH(Environ &env) {
     return BSET(env, HBit);
 }
 
-bool SRegOps::CLH(Environ &env, uint32_t instrn) {
+bool SRegOps::CLH(Environ &env, uint16_t instrn) {
     return CLH(env);
 }
 
@@ -152,7 +152,7 @@ bool SRegOps::CLH(Environ &env) {
     return BCLR(env, HBit);
 }
 
-bool SRegOps::SET(Environ &env, uint32_t instrn) {
+bool SRegOps::SET(Environ &env, uint16_t instrn) {
     return SET(env);
 }
 
@@ -160,7 +160,7 @@ bool SRegOps::SET(Environ &env) {
     return BSET(env, SBit);
 }
 
-bool SRegOps::CLT(Environ &env, uint32_t instrn) {
+bool SRegOps::CLT(Environ &env, uint16_t instrn) {
     return CLT(env);
 }
 
@@ -168,7 +168,7 @@ bool SRegOps::CLT(Environ &env) {
     return BCLR(env, TBit);
 }
 
-bool SRegOps::BST(Environ &env, uint32_t instrn) {
+bool SRegOps::BST(Environ &env, uint16_t instrn) {
     FiveBit tgtReg; ThreeBit bitNum;
     ArgsDecode::Reg5SBit3(instrn, tgtReg, bitNum);
     return BST(env, tgtReg, bitNum);
@@ -181,7 +181,7 @@ bool SRegOps::BST(Environ &env, FiveBit tgtAddr, ThreeBit regNum) {
     return false;
 }
 
-bool SRegOps::BLD(Environ &env, uint32_t instrn) {
+bool SRegOps::BLD(Environ &env, uint16_t instrn) {
     FiveBit tgtReg; ThreeBit bitNum;
     ArgsDecode::Reg5SBit3(instrn, tgtReg, bitNum);
     return BLD(env, tgtReg, bitNum);
