+++ src/simulation/sreg_ops.h
@@ -8,51 +8,51 @@
 class SRegOps {
 public:
 
-    static bool BSET(Environ& env, u_int32_t instrn);
+    static bool BSET(Environ& env, uint16_t instrn);
     static bool BSET(Environ& env, ThreeBit regNum, bool dummy);
 
-    static bool BCLR(Environ& env, u_int32_t instrn);
+    static bool BCLR(Environ& env, uint16_t instrn);
     static bool BCLR(Environ& env, ThreeBit regNum, bool dummy);
 
-    static bool SEC(Environ& env, u_int32_t instrn);
+    static bool SEC(Environ& env, uint16_t instrn);
     static bool SEC(Environ& env);
-    static bool CLC(Environ& env, u_int32_t instrn);
+    static bool CLC(Environ& env, uint16_t instrn);
     static bool CLC(Environ& env);
 
-    static bool SEZ(Environ& env, u_int32_t instrn);
+    static bool SEZ(Environ& env, uint16_t instrn);
     static bool SEZ(Environ& env);
-    static bool CLZ(Environ& env, u_int32_t instrn);
+    static bool CLZ(Environ& env, uint16_t instrn);
     static bool CLZ(Environ& env);
 
-    static bool SEN(Environ& env, u_int32_t instrn);
+    static bool SEN(Environ& env, uint16_t instrn);
     static bool SEN(Environ& env);
-    static bool CLN(Environ& env, u_int32_t instrn);
+    static bool CLN(Environ& env, uint16_t instrn);
     static bool CLN(Environ& env);
 
-    static bool SEV(Environ& env, uint32_t instrn);
+    static bool SEV(Environ& env, uint16_t instrn);
     static bool SEV(Environ& env);
-    static bool CLV(Environ& env, uint32_t instrn);
+    static bool CLV(Environ& env, uint16_t instrn);
     static bool CLV(Environ& env);
 
-    static bool SES(Environ& env, uint32_t instrn);
+    static bool SES(Environ& env, uint16_t instrn);
     static bool SES(Environ& env);
-    static bool CLS(Environ& env, uint32_t instrn);
+    static bool CLS(Environ& env, uint16_t instrn);
     static bool CLS(Environ& env);
 
-    static bool SEH(Environ& env, uint32_t instrn);
+    static bool SEH(Environ& env, uint16_t instrn);
     static bool SEH(Environ& env);
-    static bool CLH(Environ& env, uint32_t instrn);
+    static bool CLH(Environ& env, uint16_t instrn);
     static bool CLH(Environ& env);
 
-    static bool SET(Environ& env, uint32_t instrn);
+    static bool SET(Environ& env, uint16_t instrn);
     static bool SET(Environ& env);
-    static bool CLT(Environ& env, uint32_t instrn);
+    static bool CLT(Environ& env, uint16_t instrn);
     static bool CLT(Environ& env);
 
-    static bool BST(Environ& env, uint32_t instrn);
+    static bool BST(Environ& env, uint16_t instrn);
     static bool BST(Environ& env, FiveBit tgtAddr, ThreeBit regNum);
 
-    static bool BLD(Environ& env, uint32_t instrn);
+    static bool BLD(Environ& env, uint16_t instrn);
     static bool BLD(Environ& env, FiveBit tgtAddr, ThreeBit regNum);
